+++ src/main/java/com/geogym/trainer/service/TrainerService.java
@@ -193,7 +193,7 @@ public interface TrainerService {
 	 * @param trainer -  PK를 기준으로 한다
 	 * @return int - 별점 수치를 반환한다
 	 */
-	int getReputation(Trainer trainer);
+	double getReputation(Trainer trainer);
 	
 	/**
 	 * 트레이너 평가(별점) 등록
